[QA] AddOportunityWithNewTourist  and OportunityWithNewTourist  - do it!!!
