+++ src/test/java/activityTest/EditActivityTest.java
@@ -1,40 +1,38 @@
-package activity;
+package activityTest;
 
 import org.junit.After;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import pages.ParentPage;
 import parentTest.ParentTest;
 
 import java.io.IOException;
-import java.util.Set;
 
-public class EditActivityNewTab extends ParentTest {
-    final String nameOfActivity = "editActivityTestDateNewTab";
+
+public class EditActivityTest extends ParentTest {
+    final String nameOfActivity = "editActivityTestDate";
     final String valueSearchTypeDD = "2";
 
-    WebDriverWait webDriverWait20;
+
     /**
      * @throws IOException
+     * @throws InterruptedException
      * @userValidLogInExcel(_,_)
      * "login_R.O","pass_R.O"  - руководитель офиса
      * "login_Men","pass_Men"  - менеджер
-     *  valueSearchTypeDD value="1" -На сегодня
-     *                    value="2" -Будущие действия
-     *                    value="4" -Закрытые действия
-     *                    value="5" -Все по офису
-     *                    value="6" -Просроченные по офису
+     *       valueSearchTypeDD value="1" -На сегодня
+     *                         value="2" -Будущие действия
+     *                         value="4" -Закрытые действия
+     *                         value="5" -Все по офису
+     *                         value="6" -Просроченные по офису
      */
+
     @Test
-    public void editActivityNewTab() throws IOException {
+    public void editActivityTestDate() throws IOException, InterruptedException {
         //loginPage.userValidLogInExcel("login_Men", "pass_Men");
         loginPage.userValidLogInExcel("login_R.O", "pass_R.O");
         homePage.cheekCurrentUrl();
         homePage.clickOnMenuActivity();
         activityPage.cheekCurrentUrl();
-        activityPage.deletingActivityWithName(nameOfActivity, valueSearchTypeDD);
+        activityPage.deletingActivityWithName(nameOfActivity,valueSearchTypeDD);
         activityPage.deletingActivityWithName(nameOfActivity + " edit",valueSearchTypeDD);
         activityPage.clickOnButtonCreate();
         createActivityPage.cheekCurrentUrl();
@@ -42,15 +40,13 @@ public class EditActivityNewTab extends ParentTest {
         createActivityPage.clickOnDateBegin();
         createActivityPage.clickButtonActivitySave();
         editActivityPage.clickOnMenuActivity();
-        activityPage.selectSearchTypeDD("2");
-        activityPage.clickOnActivity(nameOfActivity);
-        String paretnHandle = webDriver.getWindowHandle();
-        activityPage.clickOnButtonEdit();
-
-        editActivityPage.enterActivityInfoDescription(" edit");
-        editActivityPage.clickButtonActivitySaveAndClose();
-        webDriver.switchTo().window(paretnHandle);
+        activityPage.selectSearchTypeDD(valueSearchTypeDD);
         activityPage.cheekCurrentUrl();
+        activityPage.rightClickMenu(nameOfActivity);
+        editActivityPage.enterActivityInfoDescription(" edit");
+        editActivityPage.clickButtonActivitySave();
+        activityPage.cheekCurrentUrl();
+        activityPage.selectSearchTypeDD(valueSearchTypeDD);
 
         checkAC("New Activity wasn't added",
                 activityPage.isNewActivityAdded(nameOfActivity + " edit"),
@@ -59,7 +55,9 @@ public class EditActivityNewTab extends ParentTest {
 
     @After
     public void deletingNewActivityw() {
-        activityPage.deletingActivityWithName(nameOfActivity + " edit", valueSearchTypeDD);
+        activityPage.deletingActivityWithName(nameOfActivity + " edit",valueSearchTypeDD);
 
     }
 }
+
+
